refactor: exporta função de tempo relativo para useDashboardMetrics
