+++ src/components/dashboard/RecentActivities.jsx
@@ -1,43 +1,6 @@
 import { useMemo } from 'react'
 import { Card, CardHeader, CardTitle, CardContent, CardDescription } from '@/components/ui/card.jsx'
-
-function formatarTempoRelativo(atividadesRecentes) {
-  if (!atividadesRecentes || !Array.isArray(atividadesRecentes)) return []
-
-  const agoraMs = Date.now()
-  const umMinuto = 60 * 1000
-  const umaHora = 60 * umMinuto
-  const umDia = 24 * umaHora
-  const umMes = 30 * umDia
-  const umAno = 365 * umDia
-
-  return atividadesRecentes.slice(0, 4).map((atividade) => {
-    const ms = new Date(atividade.dateTime).getTime()
-    const diferenca = agoraMs - ms
-    let tempoRelativo = ''
-
-    if (isNaN(ms) || ms > agoraMs) {
-      tempoRelativo = 'Agora mesmo'
-    } else if (diferenca < umaHora) {
-      const literal = Math.max(0, Math.floor(diferenca / umMinuto))
-      tempoRelativo = `${literal} min`
-    } else if (diferenca < umDia) {
-      const literal = Math.floor(diferenca / umaHora)
-      tempoRelativo = `${literal} ${literal > 1 ? 'horas' : 'hora'}`
-    } else if (diferenca < umMes) {
-      const literal = Math.floor(diferenca / umDia)
-      tempoRelativo = `${literal} ${literal > 1 ? 'dias' : 'dia'}`
-    } else if (diferenca < umAno) {
-      const literal = Math.floor(diferenca / umMes)
-      tempoRelativo = `${literal} ${literal > 1 ? 'meses' : 'mês'}`
-    } else {
-      const literal = Math.floor(diferenca / umAno)
-      tempoRelativo = `${literal} ${literal > 1 ? 'anos' : 'ano'}`
-    }
-
-    return tempoRelativo
-  })
-}
+import { formatarTempoRelativo } from '@/lib/useDashboardMetrics.js'
 
 export function RecentActivities({ atividadesRecentes }) {
   const atividadesFormatadas = useMemo(() => {
